fixed  bug in filemanager, where uploading image and none exists from before it cancels upload fixed buttons in manga section redirecting properly updated manga&anime index/edit/create blade's

+++ laravel_failai/app/Http/Managers/FileManager.php
@@ -57,7 +57,9 @@ class FileManager
     public function delete(array $paths)
     {
         foreach($paths as $path) {
-            Storage::delete($path);
+            if ($path !== null) {
+                Storage::delete($path);
+            }
         }
     }
 }
